+++ app/src/main/java/mega/privacy/android/app/MegaApplication.java
@@ -48,6 +48,7 @@ import mega.privacy.android.app.lollipop.ManagerActivityLollipop;
 import mega.privacy.android.app.lollipop.controllers.AccountController;
 import mega.privacy.android.app.lollipop.megachat.ChatItemPreferences;
 import mega.privacy.android.app.lollipop.megachat.ChatSettings;
+import mega.privacy.android.app.lollipop.megachat.NotificationBuilder;
 import mega.privacy.android.app.lollipop.megachat.RecentChatsFragmentLollipop;
 import mega.privacy.android.app.lollipop.megachat.chatAdapters.MegaChatLollipopAdapter;
 import mega.privacy.android.app.lollipop.megachat.chatAdapters.MegaListChatLollipopAdapter;
@@ -84,7 +85,7 @@ public class MegaApplication extends Application implements MegaListenerInterfac
 
 
 	MegaChatApiAndroid megaChatApi = null;
-
+	private NotificationBuilder notificationBuilder;
 //	static final String GA_PROPERTY_ID = "UA-59254318-1";
 //	
 //	/**
@@ -254,6 +255,8 @@ public class MegaApplication extends Application implements MegaListenerInterfac
 				MegaChatApiAndroid.setLogLevel(MegaChatApiAndroid.LOG_LEVEL_ERROR);
 			}
 		}
+
+		notificationBuilder =  NotificationBuilder.newInstance(this);
 		
 //		initializeGA();
 		
@@ -679,6 +682,9 @@ public class MegaApplication extends Application implements MegaListenerInterfac
 		log("showNotification: "+item.getTitle()+ "message: "+item.getLastMessage());
 
 		ChatSettings chatSettings = dbH.getChatSettings();
+		String email = megaChatApi.getContactEmail(item.getPeerHandle());
+
+		String color = megaApi.getUserAvatarColor(MegaApiAndroid.userHandleToBase64(item.getPeerHandle()));
 		if(chatSettings!=null){
 			if(chatSettings.getNotificationsEnabled().equals("true")){
 				log("Notifications ON for all chats");
@@ -694,20 +700,20 @@ public class MegaApplication extends Application implements MegaListenerInterfac
 					if(soundString.equals("true")||soundString.equals("")){
 
 						Uri defaultSoundUri2 = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION);
-						showCustomizedNotification(defaultSoundUri2, item, chatSettings.getVibrationEnabled());
+						notificationBuilder.sendBundledNotification(defaultSoundUri2, item, chatSettings.getVibrationEnabled(), email, color);
 					}
 					else if(soundString.equals("-1")){
 						log("Silent notification");
-						showCustomizedNotification(null, item, chatSettings.getVibrationEnabled());
+						notificationBuilder.sendBundledNotification(null, item, chatSettings.getVibrationEnabled(), email, color);
 					}
 					else{
 						Ringtone sound = RingtoneManager.getRingtone(this, uri);
 						if(sound==null){
 							log("Sound is null");
-							showCustomizedNotification(null, item, chatSettings.getVibrationEnabled());
+							notificationBuilder.sendBundledNotification(null, item, chatSettings.getVibrationEnabled(), email, color);
 						}
 						else{
-							showCustomizedNotification(uri, item, chatSettings.getVibrationEnabled());
+							notificationBuilder.sendBundledNotification(uri, item, chatSettings.getVibrationEnabled(), email, color);
 						}
 					}
 				}
@@ -722,20 +728,21 @@ public class MegaApplication extends Application implements MegaListenerInterfac
 						if(soundString.equals("true")){
 
 							Uri defaultSoundUri2 = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION);
-							showCustomizedNotification(defaultSoundUri2, item, chatSettings.getVibrationEnabled());
+							notificationBuilder.sendBundledNotification(defaultSoundUri2, item, chatSettings.getVibrationEnabled(), email, color);
 						}
 						else if(soundString.equals("-1")){
 							log("Silent notification");
-							showCustomizedNotification(null, item, chatSettings.getVibrationEnabled());
+							notificationBuilder.sendBundledNotification(null, item, chatSettings.getVibrationEnabled(), email, color);
 						}
 						else{
 							Ringtone sound = RingtoneManager.getRingtone(this, uri);
 							if(sound==null){
 								log("Sound is null");
-								showCustomizedNotification(null, item, chatSettings.getVibrationEnabled());
+								notificationBuilder.sendBundledNotification(null, item, chatSettings.getVibrationEnabled(), email, color);
 							}
 							else{
-								showCustomizedNotification(uri, item, chatSettings.getVibrationEnabled());
+								notificationBuilder.sendBundledNotification(uri, item, chatSettings.getVibrationEnabled(), email, color);
+
 							}
 						}
 					}
@@ -752,265 +759,9 @@ public class MegaApplication extends Application implements MegaListenerInterfac
 			log("Notifications DEFAULT ON");
 
 			Uri defaultSoundUri2 = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION);
-			showCustomizedNotification(defaultSoundUri2, item, "true");
+			notificationBuilder.sendBundledNotification(defaultSoundUri2, item, "true", email, color);
 		}
 	}
-
-	public void showCustomizedNotification(Uri uriParameter, MegaChatListItem item, String vibration){
-		log("showCustomizedNotification");
-		String notificationTitle = "Chat activity";
-		String source = "<b>"+item.getTitle()+": </b>"+item.getLastMessage();
-
-		Spanned notificationContent = Html.fromHtml(source,0);
-
-		int notificationId = Constants.NOTIFICATION_PUSH_CHAT;
-
-		Intent intent = new Intent(this, ManagerActivityLollipop.class);
-		intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
-		intent.setAction(Constants.ACTION_CHAT_NOTIFICATION_MESSAGE);
-		PendingIntent pendingIntent = PendingIntent.getActivity(this, 0 /* Request code */, intent,
-				PendingIntent.FLAG_ONE_SHOT);
-
-
-		Uri defaultSoundUri = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION);
-		NotificationCompat.Builder notificationBuilder = new NotificationCompat.Builder(this)
-				.setSmallIcon(R.drawable.ic_stat_notify_download)
-				.setContentTitle(notificationTitle)
-				.setContentText(notificationContent)
-				.setColor(ContextCompat.getColor(getApplicationContext(),R.color.mega))
-				.setAutoCancel(true)
-				.setSound(defaultSoundUri)
-				.setContentIntent(pendingIntent);
-
-		Bitmap largeIcon = setUserAvatar(item);
-		if(largeIcon!=null){
-			log("There is avatar!");
-			notificationBuilder.setLargeIcon(largeIcon);
-		}
-
-		NotificationCompat.InboxStyle inboxStyle = new NotificationCompat.InboxStyle();
-
-//		StringBuilder[] events = {notificationContent, new StringBuilder("y trooo"), new StringBuilder("y moreee"), new StringBuilder("y yaaaa")};
-// Sets a title for the Inbox in expanded layout
-//		inboxStyle.setBigContentTitle("New messages:");
-
-		String[] events = {"y trooo", "y moreee", "y yaaaa"};
-// Moves events into the expanded layout
-		inboxStyle.addLine(notificationContent);
-		for (int i=0; i < events.length; i++) {
-			inboxStyle.addLine(events[i]);
-
-		}
-// Moves the expanded layout object into the notification object.
-		notificationBuilder.setStyle(inboxStyle);
-
-
-		NotificationManager notificationManager =
-				(NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
-		notificationBuilder.setSound(uriParameter);
-		if(vibration!=null){
-			if(vibration.equals("true")){
-				notificationBuilder.setVibrate(new long[] {0, 1000});
-			}
-		}
-
-		notificationManager.notify(notificationId, notificationBuilder.build());
-	}
-
-	public Bitmap setUserAvatar(MegaChatListItem item){
-		log("setUserAvatar");
-
-		if(item.isGroup()){
-			return createGroupChatAvatar(item);
-		}
-		else{
-			if(megaChatApi==null){
-				return createDefaultAvatar(item);
-			}
-			String contactMail = megaChatApi.getContactEmail(item.getPeerHandle());
-			File avatar = null;
-			if (getExternalCacheDir() != null){
-				avatar = new File(getExternalCacheDir().getAbsolutePath(), contactMail + ".jpg");
-			}
-			else{
-				avatar = new File(getCacheDir().getAbsolutePath(), contactMail + ".jpg");
-			}
-			Bitmap bitmap = null;
-			if (avatar.exists()){
-				if (avatar.length() > 0){
-					BitmapFactory.Options bOpts = new BitmapFactory.Options();
-					bOpts.inPurgeable = true;
-					bOpts.inInputShareable = true;
-					bitmap = BitmapFactory.decodeFile(avatar.getAbsolutePath(), bOpts);
-					if (bitmap == null) {
-						return createDefaultAvatar(item);
-					}
-					else{
-						return getCircleBitmap(bitmap);
-					}
-				}
-				else{
-					return createDefaultAvatar(item);
-				}
-			}
-			else{
-				return createDefaultAvatar(item);
-			}
-
-		}
-	}
-
-	private Bitmap getCircleBitmap(Bitmap bitmap) {
-		final Bitmap output = Bitmap.createBitmap(bitmap.getWidth(),
-				bitmap.getHeight(), Bitmap.Config.ARGB_8888);
-		final Canvas canvas = new Canvas(output);
-
-		final int color = Color.RED;
-		final Paint paint = new Paint();
-		final Rect rect = new Rect(0, 0, bitmap.getWidth(), bitmap.getHeight());
-		final RectF rectF = new RectF(rect);
-
-		paint.setAntiAlias(true);
-		canvas.drawARGB(0, 0, 0, 0);
-		paint.setColor(color);
-		canvas.drawOval(rectF, paint);
-
-		paint.setXfermode(new PorterDuffXfermode(PorterDuff.Mode.SRC_IN));
-		canvas.drawBitmap(bitmap, rect, rect, paint);
-
-		bitmap.recycle();
-
-		return output;
-	}
-
-	public Bitmap createDefaultAvatar(MegaChatListItem item){
-		log("createDefaultAvatar()");
-
-		Bitmap defaultAvatar = Bitmap.createBitmap(Constants.DEFAULT_AVATAR_WIDTH_HEIGHT,Constants.DEFAULT_AVATAR_WIDTH_HEIGHT, Bitmap.Config.ARGB_8888);
-		Canvas c = new Canvas(defaultAvatar);
-		Paint p = new Paint();
-		p.setAntiAlias(true);
-
-		String color = megaApi.getUserAvatarColor(MegaApiAndroid.userHandleToBase64(item.getPeerHandle()));
-		if(color!=null){
-			log("The color to set the avatar is "+color);
-			p.setColor(Color.parseColor(color));
-		}
-		else{
-			log("Default color to the avatar");
-			p.setColor(getResources().getColor(R.color.lollipop_primary_color));
-		}
-
-		int radius;
-		if (defaultAvatar.getWidth() < defaultAvatar.getHeight())
-			radius = defaultAvatar.getWidth()/2;
-		else
-			radius = defaultAvatar.getHeight()/2;
-
-		c.drawCircle(defaultAvatar.getWidth()/2, defaultAvatar.getHeight()/2, radius, p);
-		p.setColor(getResources().getColor(R.color.white));
-		c.drawText("P", 10, 10, p);
-
-		return defaultAvatar;
-//		holder.imageView.setImageBitmap(defaultAvatar);
-//
-//		boolean setInitialByMail = false;
-//
-//		if (holder.fullName != null){
-//			if (holder.fullName.trim().length() > 0){
-//				String firstLetter = holder.fullName.charAt(0) + "";
-//				firstLetter = firstLetter.toUpperCase(Locale.getDefault());
-//				holder.contactInitialLetter.setText(firstLetter);
-//				holder.contactInitialLetter.setTextColor(Color.WHITE);
-//				holder.contactInitialLetter.setVisibility(View.VISIBLE);
-//			}else{
-//				setInitialByMail=true;
-//			}
-//		}
-//		else{
-//			setInitialByMail=true;
-//		}
-//		if(setInitialByMail){
-//			if (holder.contactMail != null){
-//				if (holder.contactMail.length() > 0){
-//					log("email TEXT: " + holder.contactMail);
-//					log("email TEXT AT 0: " + holder.contactMail.charAt(0));
-//					String firstLetter = holder.contactMail.charAt(0) + "";
-//					firstLetter = firstLetter.toUpperCase(Locale.getDefault());
-//					holder.contactInitialLetter.setText(firstLetter);
-//					holder.contactInitialLetter.setTextColor(Color.WHITE);
-//					holder.contactInitialLetter.setVisibility(View.VISIBLE);
-//				}
-//			}
-//		}
-//		holder.contactInitialLetter.setTextSize(24);
-	}
-
-	public Bitmap createGroupChatAvatar(MegaChatListItem item){
-		log("createGroupChatAvatar()");
-
-		Bitmap defaultAvatar = Bitmap.createBitmap(Constants.DEFAULT_AVATAR_WIDTH_HEIGHT,Constants.DEFAULT_AVATAR_WIDTH_HEIGHT, Bitmap.Config.ARGB_8888);
-		Canvas c = new Canvas(defaultAvatar);
-		Paint p = new Paint();
-		p.setAntiAlias(true);
-		p.setColor(ContextCompat.getColor(this,R.color.divider_upgrade_account));
-
-		int radius;
-		if (defaultAvatar.getWidth() < defaultAvatar.getHeight())
-			radius = defaultAvatar.getWidth()/2;
-		else
-			radius = defaultAvatar.getHeight()/2;
-
-		c.drawCircle(defaultAvatar.getWidth()/2, defaultAvatar.getHeight()/2, radius, p);
-
-		if(item.getTitle()!=null){
-			if(!item.getTitle().isEmpty()){
-				char title = item.getTitle().charAt(0);
-				String firstLetter = new String(title+"");
-
-				if(!firstLetter.equals("(")){
-
-                    log("Draw letter: "+firstLetter);
-					Paint text = new Paint();
-					Typeface face = Typeface.SANS_SERIF;
-					text.setTypeface(face);
-					text.setAntiAlias(true);
-					text.setSubpixelText(true);
-					text.setStyle(Paint.Style.FILL);
-					text.setColor(Color.WHITE);
-					text.setTextSize(150);
-					text.setTextAlign(Paint.Align.CENTER);
-
-					text.setColor(ContextCompat.getColor(this,R.color.black));
-                    Rect bounds = new Rect();
-					text.getTextBounds(firstLetter, 0, firstLetter.length(), bounds);
-                    int x = (defaultAvatar.getWidth() - bounds.width())/2;
-                    int y = (defaultAvatar.getHeight() + bounds.height())/2;
-                    c.drawText(firstLetter.toUpperCase(Locale.getDefault()), x, y, text);
-				}
-
-			}
-		}
-		return defaultAvatar;
-//		String firstLetter = holder.contactInitialLetter.getText().toString();
-//
-//		if(firstLetter.trim().isEmpty()){
-//			holder.contactInitialLetter.setVisibility(View.INVISIBLE);
-//		}
-//		else{
-//			log("Group chat initial letter is: "+firstLetter);
-//			if(firstLetter.equals("(")){
-//				holder.contactInitialLetter.setVisibility(View.INVISIBLE);
-//			}
-//			else{
-//				holder.contactInitialLetter.setText(firstLetter);
-//				holder.contactInitialLetter.setTextColor(Color.WHITE);
-//				holder.contactInitialLetter.setVisibility(View.VISIBLE);
-//				holder.contactInitialLetter.setTextSize(24);
-//			}
-//		}
-	}
-
 
 	@Override
 	public void onChatInitStateUpdate(MegaChatApiJava api, int newState) {
